Fix circular dependency
Fix circluar dependency introduces by using type hints
