+++ relay/api/exchange/resources.py
@@ -4,7 +4,7 @@ from webargs import fields as webfields
 from webargs.flaskparser import abort
 from eth_utils import to_checksum_address
 
-from relay.main import TrustlinesRelay
+from relay.relay import TrustlinesRelay
 from relay.api import fields
 from relay.exchange.order import Order
 from relay.exchange.orderbook import OrderInvalidException
