+++ relay/api/messaging/resources.py
@@ -4,7 +4,7 @@ from flask_restful import Resource
 from webargs import fields
 from webargs.flaskparser import use_args
 
-from relay.main import TrustlinesRelay
+from relay.relay import TrustlinesRelay
 
 
 class PostMessage(Resource):
